+++ src/js/content/cellar/tiles/ziggurat.js
@@ -1,12 +1,12 @@
 content.cellar.tiles.ziggurat = content.cellar.tiles.invent({
   id: 'ziggurat',
   name: 'The ziggurat',
+  category: 'situational',
   uniquePerFloor: true,
   weight: 2,
   defaultState: {
     delta: 0,
   },
-  canGenerate: (tile) => engine.fn.between(tile.z, -1, -2), // Not on first floor
   onEnterEffects: function () {
     // Health trends toward 50% max